% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosleepr.R
\name{autosleep_sleep_df}
\alias{autosleep_sleep_df}
\title{Autosleep dataframe}
\usage{
autosleep_sleep_df(pathname = getwd())
}
\arguments{
\item{pathname}{path to the raw Autosleep.csv file}
}
\description{
Return dataframe for all sleep data
}
