% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosleepr.R
\name{autosleep_sleep_raw}
\alias{autosleep_sleep_raw}
\title{Raw Autosleep dataframe}
\usage{
autosleep_sleep_raw(pathname = getwd())
}
\arguments{
\item{pathname}{path to the raw Autosleep.csv file}
}
\description{
Return Autosleep dataframe in raw form
}
