% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronometer.R
\name{cronometer_df}
\alias{cronometer_df}
\title{Cronometer dataframe}
\usage{
cronometer_df(pathname = getwd())
}
\arguments{
\item{pathname}{path to the raw Cronometer.csv file}
}
\description{
Return dataframe for all cronometer data
}
