% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_tracker.R
\name{inside_tracker_df}
\alias{inside_tracker_df}
\title{Return a canonical Inside Tracker blood test results dataframe}
\usage{
inside_tracker_df(
  filename = system.file("extdata", package = "qselfr", "InsideTrackerSample.csv")
)
}
\arguments{
\item{filename}{path to an Inside Tracker CSV file}
}
\value{
dataframe
}
\description{
Return a canonical Inside Tracker blood test results dataframe
}
