% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_tracker.R
\name{inside_tracker_plot}
\alias{inside_tracker_plot}
\title{Pretty plot of Inside Tracker dataframe}
\usage{
inside_tracker_plot(df)
}
\arguments{
\item{dataframe}{canonical inside tracker dataframe}
}
\description{
Pretty plot of Inside Tracker dataframe
}
