% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keto_mojo.R
\name{keto_mojo_df}
\alias{keto_mojo_df}
\title{Read a Keto Mojo CSV file as a dataframe}
\usage{
keto_mojo_df(filename)
}
\arguments{
\item{filename}{path to a properly formatted Keto Mojo CSV file.}
}
\value{
dataframe
}
\description{
Read a Keto Mojo CSV file as a dataframe
}
